Cleanup and workflow support
Since Rosemary works through one client at a time, we will not make a
minimum client requirement.  That way, only the clients she's added will
she be prompted for info for.  This could result in a single party
regular mediation, but honestly, the other party isn't likely to be
missed and probably doesn't require a prompt.